Weather App
This Weather App fetches and displays real-time weather information for a specified city using the OpenWeatherMap API. To use the Weather App, users enter a city name into the input field and click the search button. The app then fetches real-time weather data from the OpenWeatherMap API. The JavaScript function checkWeather(city) sends a request to the API and handles the response. If the city is found, the app updates the webpage with the current temperature, humidity, wind speed, and a corresponding weather icon. If the city is not found, an error message is displayed instead.